More name fixing.
