+++ router/underlayproviders/afpacketudpip/BUILD.bazel
@@ -5,26 +5,28 @@ load("//tools/lint:go.bzl", "go_library")
 genrule(
     name = "gen_bpf_filter_go",
     srcs = [
-        "bpf_filter.c",
+        "portfilter.c",
         "bpf_helpers.h",
         "bpf_helper_defs.h",
     ],
     outs = [
-        "bpf_filter_bpfel.go",
-        "bpf_filter_bpfel.o",
+        "portfilter_bpfel.go",
+        "portfilter_bpfel.o",
     ],
-    cmd = "GOPACKAGE=af_packet_udpip $(execpath @com_github_cilium_ebpf//cmd/bpf2go) -output-dir $$(dirname $(location bpf_filter_bpfel.go)) -tags linux bpf_filter $(location bpf_filter.c)",
+    cmd = "GOPACKAGE=afpacketudpip $(execpath @com_github_cilium_ebpf//cmd/bpf2go) -output-dir $$(dirname $(location portfilter_bpfel.go)) -tags linux portfilter $(location portfilter.c)",
     tools = ["@com_github_cilium_ebpf//cmd/bpf2go"],
 )
 
 go_library(
     name = "go_default_library",
     srcs = [
-        "bpf_filter_bpfel.go",
-        "usermode_filter.go",
+        "bpf_helper_defs.h",
+        "bpf_helpers.h",
+        "portfilter_bpfel.go",
+        "rawsocket.go",
     ],
     embedsrcs = [
-        "bpf_filter_bpfel.o",  #keep
+        "portfilter_bpfel.o",  #keep
     ],
     importpath = "github.com/scionproto/scion/router/underlayproviders/afpacketudpip",
     visibility = ["//visibility:public"],
